LPS-89415 generator should only delete files with @generated in them

+++ modules/util/portal-tools-rest-builder/src/main/java/com/liferay/portal/tools/rest/builder/internal/util/FileUtil.java
@@ -59,6 +59,12 @@ public class FileUtil {
 						return FileVisitResult.CONTINUE;
 					}
 
+					String content = read(file);
+
+					if (!content.contains("@generated")) {
+						return FileVisitResult.CONTINUE;
+					}
+
 					Files.delete(path);
 
 					System.out.println("Deleting " + file.getCanonicalPath());
